requirements.txt
# Chess Engine Testing Framework - Python Dependencies

# Core Chess Library
python-chess==1.999

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Frontend Template Engine
jinja2==3.1.2

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Database
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
